Just a start
NOTE: The "mainwindow" is the splash screen that starts up. Since that
can't be closed without closing the other windows, I can only hide it
(which causes a memory leak since it is never deleted). However, I
couldn't think of any other way to control whether the gamewindow or
mapcreator should be shown. We can discuss this when we meet.