Container Builder
The code appears to work I'm not 100% convinced it's doing what it's
supposed to. Regardless, this is what I submitted, and I'll fix issues
as they come up.

The files posted are the files that were changed as a result of
assignment 3. The rest of the project was untouched and can stay that
way as far as this code is concerned.